requirements.txt
pillow==10.2.0
aiogram==3.4.1
python-dotenv==1.0.1
gspread==6.0.0
yookassa==3.0.1
